Replaced the dict parameter in Payload.
Instead of passing a dictionary, changed to passing each parameter separately.
This reduces the overhead of having to deal with a dict for only five parameters.
Based on this, I also removed the package building function in sensor.py since it became useless after this change. When data is collected from the sensors directly, it is stored in a Payload object without having to pass through yet another intermediary function which would have build the Payload object instead.