Replaced data collecting with dict to list.
Instead of collecting data in a dictionary of the following structure:

{'temperature': Payload(...),
 'altitude': Payload(...),
...},

Used lists instead to only store the Payload objects for each collected packet of data. Since the information of what type it is is already existing in the Payload, there is no need to store it as a dictionary key and therefore add overhead.

Current structure:

[ Payload('temperature', ...), Payload('altitude', ...), ... ].